Restructure the database test modules
