+++ database/mysql/src/test/java/org/ballerinalang/scenario/test/database/mysql/Constants.java
@@ -1,0 +1,48 @@
+package org.ballerinalang.scenario.test.database.mysql;
+
+public class Constants {
+    public static final String MYSQL_JDBC_URL_KEY = "database.mysql.test.jdbc.url";
+    public static final String MYSQL_JDBC_USERNAME_KEY = "database.mysql.test.jdbc.username";
+    public static final String MYSQL_JDBC_PASSWORD_KEY = "database.mysql.test.jdbc.password";
+
+    public static String BIT_VAL_FIELD = "bitVal";
+    public static String TINYINT_VAL_FIELD = "tinyIntVal";
+    public static String SMALLINT_VAL_FIELD = "smallIntVal";
+    public static String MEDIUMINT_VAL_FIELD = "mediumIntVal";
+    public static String INT_VAL_FIELD = "intVal";
+    public static String BIGINT_VAL_FIELD = "bigIntVal";
+    public static String DECIMAL_VAL_FIELD = "decimalVal";
+    public static String NUMERIC_VAL_FIELD = "numericVal";
+
+    public static String FLOAT_VAL_FIELD = "floatVal";
+    public static String DOUBLE_VAL_FIELD = "doubleVal";
+
+    public static String CHAR_VAL_FIELD = "charVal";
+    public static String VARCHAR_VAL_FIELD = "varcharVal";
+    public static String TINYTEXT_VAL_FIELD = "tinyTextVal";
+    public static String TEXT_VAL_FIELD = "textVal";
+    public static String MEDIUMTEXT_VAL_FIELD = "mediumTextVal";
+    public static String LONGTEXT_VAL_FIELD = "longTextVal";
+    public static String SET_VAL_FIELD = "setVal";
+    public static String ENUM_VAL_FIELD = "enumVal";
+
+    public static String DATE_VAL_FIELD_INT = "dateInt";
+    public static String TIME_VAL_FIELD_INT = "timeInt";
+    public static String DATETIME_VAL_FIELD_INT = "dateTimeInt";
+    public static String TIMESTAMP_VAL_FIELD_INT = "timestampInt";
+    public static String YEAR_VAL_FIELD_INT = "yearInt";
+
+    public static String DATE_VAL_FIELD_REC = "dateRec";
+    public static String TIME_VAL_FIELD_REC = "timeRec";
+    public static String DATETIME_VAL_FIELD_REC = "dateTimeRec";
+    public static String TIMESTAMP_VAL_FIELD_REC = "timestampRec";
+    public static String YEAR_VAL_FIELD_REC = "yearRec";
+
+    public static String BINARY_FIELD = "binaryVal";
+    public static String VARBINARY_FIELD = "varBinaryVal";
+    public static String TINYBLOB_FIELD = "tinyBlobVal";
+    public static String BLOB_FIELD = "blobVal";
+    public static String MEDIUMBLOB_FIELD = "mediumBlobVal";
+    public static String LONGBLOB_FIELD = "longBlobVal";
+
+}
