+++ database/mysql/src/test/resources/bal-src/batch-update-test.bal
