+++ database/mysql/src/test/resources/sql-src/cleanup-call-test.sql
