+++ database/mysql/src/test/resources/sql-src/cleanup-select-update-test.sql
