+++ database/mysql/src/test/resources/sql-src/ddl-select-update-test.sql
