+++ database/mysql/src/test/resources/sql-src/ddl-test.sql
