+++ database/postgresql/src/test/resources/sql-src/cleanup-ddl-test.sql
