+++ database/postgresql/src/test/resources/sql-src/cleanup-select-test.sql
