+++ pom.xml
@@ -91,7 +91,6 @@
                 <groupId>org.ballerinalang</groupId>
                 <artifactId>ballerina-core</artifactId>
                 <version>${ballerina.version}</version>
-                <scope>test</scope>
             </dependency>
             <dependency>
                 <groupId>org.ballerinalang</groupId>
@@ -227,8 +226,48 @@
                 <scope>test</scope>
                 <version>${postgres.driver.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.ballerinalang</groupId>
+                <artifactId>util</artifactId>
+                <version>${project.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
+
+    <repositories>
+        <repository>
+            <id>wso2.releases</id>
+            <name>WSO2 Releases Repository</name>
+            <url>http://maven.wso2.org/nexus/content/repositories/releases/</url>
+            <releases>
+                <enabled>true</enabled>
+                <updatePolicy>daily</updatePolicy>
+                <checksumPolicy>ignore</checksumPolicy>
+            </releases>
+        </repository>
+        <repository>
+            <id>wso2.snapshots</id>
+            <name>WSO2 Snapshot Repository</name>
+            <url>http://maven.wso2.org/nexus/content/repositories/snapshots/</url>
+            <snapshots>
+                <enabled>true</enabled>
+                <updatePolicy>daily</updatePolicy>
+            </snapshots>
+            <releases>
+                <enabled>false</enabled>
+            </releases>
+        </repository>
+        <repository>
+            <id>wso2-nexus</id>
+            <name>WSO2 internal Repository</name>
+            <url>http://maven.wso2.org/nexus/content/groups/wso2-public/</url>
+            <releases>
+                <enabled>true</enabled>
+                <updatePolicy>daily</updatePolicy>
+                <checksumPolicy>ignore</checksumPolicy>
+            </releases>
+        </repository>
+    </repositories>
 
     <profiles>
         <profile>
@@ -243,7 +282,9 @@
                 <module>bbg/kafka</module>
                 <module>bbg/bbg-integration-patterns</module>
                 <module>connectors/soap</module>
-                <module>database</module>
+                <module>database/mysql</module>
+                <module>database/postgresql</module>
+                <module>database/util</module>
             </modules>
         </profile>
         <profile>
@@ -281,7 +322,9 @@
             <id>database</id>
             <modules>
                 <module>scenarios-commons</module>
-                <module>database</module>
+                <module>database/util</module>
+                <module>database/mysql</module>
+                <module>database/postgresql</module>
             </modules>
         </profile>
     </profiles>
